Implement infinite scrolling
In the method fetchCards() now you want to pass 2 parameters to the page and size. If these parameters are not specified, the default page is 0, and the size is 28. This is the minimum size in order to fill the entire screen with cards and get the ability to scroll the page to load the rest of the cards